Added the hackbytes dependencies and resources repos as submodules
The hackbytes dependencies repository contains the necessary software
for successfully building and deploying hackbytes.com. It contains the
necessary ruby gems, binaries and other software as well as the
necessary licenses for all programs. Sources for the programs are also
provided if the terms of the license requires it.

The hackbytes resources repository contains all image assets and any
media files that hackbytes.com uses. These do not include the xcf files
from which these images can be derived. Project files are currently
managed externally, outside the scope of git repositories.

Signed-off-by: Byron Sanchez <bfs5030@gmail.com>