Added ruby and ruby-related files, as well as files used during build.
These files include scripts for building the website, as well as extra
functionality for the users of the website. The Gemfile also helps
ensure the necessary gems are installed.

In addition, scss files help with making CSS easier to organize. Though
the full power of Sass has not yet been leveraged by generalizing CSS
definitions or using other available functionality.

Signed-off-by: Byron Sanchez <bfs5030@gmail.com>