Added common project-description files.
These include copyright and licensing information as well as
instruction on how to setup a working installation.

The COPYRIGHT file describes the licensing of various components of the
repository. Images and media should be useable outside the scope of
forks, thus, they typically are licensed under a CC license of some
sort.

Unfortunately, licensing is a complex topic. There may have to be future
changes to account for situations I have not considered.

Signed-off-by: Byron Sanchez <bfs5030@gmail.com>