Added first version of table-header formatter code to markdownifier.
In order to implement responsive tables, table headers and table
columns need to have a matching axis attribute. This code scans
a rendered html file (the markdownifier output) and scans for any
tables that contain headers. The header values are then copied and
stored in the axis attributes for the corresponding header and column
values.

The axis attribute is later used to implement a responsive version of
tables where the header information is always next to the column values.
In other words, in small screens the table becomes a list.

Signed-off-by: Byron Sanchez <bfs5030@gmail.com>