Added the nuke script for facilitating removal of unlicensed material.
The nuke script allows other developers to remove material that I hold
the copyright to and that I have not licensed. In terms of the fork,
they simply act as placeholder images until the fork's developer creates
their own assets.

Nuke is helpful because the alternative would be manually sifting
through files and directories and removing material that way.

Signed-off-by: Byron Sanchez <bfs5030@gmail.com>