Updated ruby assets scripts.
Added custom tags to be used during the markdownifying process (the
video tag).

Updated the env.rb script. When the application is in development mode,
future posts will be compiled in the final output. When the application
is in production mode, future posts will NOT be compiled in the final
output.

Signed-off-by: Byron Sanchez <bfs5030@gmail.com>