Updated the compilation components to make gem installation configurable
Previously, the compilation component would package all gems on the
machine in which the compile was performed. This means that any C
extensions would be compiled on that machine. But there is no guarantee
that the machine running the "rake compile" is the main production
server (this seperation is provided by "rake deploy").

Thus, whenever the "rake env" specifies 0 (for development mode), the
"rake compile" process will package and install gems on the machine in
which the command was executed. If "rake env" specifies 1 (for
production mode), the "rake compile" process will NOT package the gems
locally. In the latter case, "rake deploy" will invoke the gem packaging
on the production server.

Signed-off-by: Byron Sanchez <bfs5030@gmail.com>