Fixed responsive design bug which kept floated the portfolio sidebar
The sidebar should no longer float for windows 767px or smaller.
Instead, the sidebar should become fluid and have a width that fills the
entire window. This particular bug made it continue to float for all
sizes below 767px which caused an unnatural appearance in the overall
design of the page.

Signed-off-by: Byron Sanchez <bfs5030@gmail.com>