Added the url alias checking system along with a custom 404 page
If a url cannot be served directly (the file doesn't exist), a lookup is
performed in a SQLite file containing all available aliases for urls. If
a matching alias is not found, the 404 page is returned. However, if an
alias is found, the page the alias points to is served.

This helps in ensuring that any url changes will never truly break
links. At least as much as such a guarantee is possible.

Signed-off-by: Byron Sanchez <bfs5030@gmail.com>