Isolated "precompilation" in build scripts
Some files cannot be generated on the server, so these items will need
to be compiled prior to deployment on the development workstation. Thus,
a convenience task for precompilation has been added, called "rake
precompile."

In production mode, precompile functions are not called under "rake
build" task since a script will automate the "rake precompile" task
invocation at the appropriate time. In development mode, precompile
functions are called when "rake build" is invoked, as a matter of
convenience.

Signed-off-by: Byron Sanchez <bfs5030@gmail.com>