Fix nodemon rules to prevent infinite refresh loop
The rules with the wildcard worked on Windows. But on Linux, the ignore
rules /did not/ take precedence over the watch rules, and thus files
that were supposed to be ignored caused nodemon to refresh in an endless
cycle.

This patch removes the wildcards, which fixes the issue, and builds
should run as expected.