Add loading of bold Ubuntu Mono fonts
Bold might be used in syntax highlighting. It might not. It is in
git.hackbytes.io, which revealed to me that forcing bold on non-bold
loaded fonts causes text to look bad. So I'm loading it here now to
prevent that issue from surfacing in the future on blog instances.