Remove unnecessary bind-mounts to reduce build-time anomalies
Builds are failing due to node_modules not properly being loaded for the
wintersmith package (specifically, coffeescript isn't being found).

My suspicion is that the blogs-libs package directory, which should
install the coffeescript module, is having it's node_modules shadowed by
the host node_modules due to a bind mount. It means the packages get
installed within the container, but the bindmount then overlays the
hosts node_modules, which is empty.

The reason we're bind-mounting at all is because the value of the web
containers are to provide a build environment for the static site. We
only need them to compile the site, and then we need the build host to
access the compiled output and sync that to the deploy target.

So I'm removing all bind-mounts that aren't needed to accomplish this
task from the prod docker config. This will reduce the anomaly vector
for this sort of thing.

This /does/ imply that the bug exists in development as well. Since the
wintersmith project structure isn't set up for easy
bind-mounting (source lives as siblings to package.json, which means we
can't cache layer node_modules without significantly diverging from
upstream project structure), I'll live with this for now until I decide
what I want to do. Maybe I won't even want to track wintersmith as a
dependency anymore, since the only reason I introduced it was to modify
it for multiple blog instances, which I didn't even have to do.

Whatevs, we'll see.