Remove overlay pattern for docker containers
The overlay pattern was something I was building when combining my
individual blog repos into one mono-repo with a single Wintersmith
engine.

The way it would work is that global files shared by all blogs would be
copied to a target directory in a docker container, and then the blog
instances would be merged with that target directory, so that I could
trivially "overlay" blog-instance specific files whenever I wanted.

Decided against this. Changes directory structure way too much and it
doesn't map to the directory structure you see when editing the code.
It's more of a design mess and hard to visualize.